line_bot/requirements.txt
line-bot-sdk==3.7.0
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
google-generativeai==0.3.2
pandas==2.2.1
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2024.1 